+++ lib/leftovers/matcher_builders/node.rb
@@ -35,6 +35,18 @@ module Leftovers
           )
         end
 
+        def build_all_matcher(all)
+          return unless all
+
+          ::Leftovers::MatcherBuilders::And.build(
+            all.map { |pattern| ::Leftovers::MatcherBuilders::Node.build(pattern) }
+          )
+        end
+
+        def build_any_matcher(any)
+          ::Leftovers::MatcherBuilders::Node.build(any)
+        end
+
         def build_from_hash( # rubocop:disable Metrics/ParameterLists
           names: nil, match: nil, has_prefix: nil, has_suffix: nil,
           document: false,
@@ -43,7 +55,7 @@ module Leftovers
           has_receiver: nil,
           type: nil,
           privacy: nil,
-          unless_arg: nil
+          unless_arg: nil, all: nil, any: nil
         )
           ::Leftovers::MatcherBuilders::And.build([
             build_node_name_matcher(names, match, has_prefix, has_suffix),
@@ -53,7 +65,7 @@ module Leftovers
             ::Leftovers::MatcherBuilders::NodeHasReceiver.build(has_receiver),
             ::Leftovers::MatcherBuilders::NodePrivacy.build(privacy),
             ::Leftovers::MatcherBuilders::NodeType.build(type),
-            build_unless_matcher(unless_arg)
+            build_unless_matcher(unless_arg), build_all_matcher(all), build_any_matcher(any)
           ])
         end
       end