+++ lib/leftovers/processor_builders/argument.rb
@@ -3,12 +3,14 @@
 module Leftovers
   module ProcessorBuilders
     module Argument
-      def self.build(patterns, processor)
+      def self.build(patterns, processor) # rubocop:disable Metrics/MethodLength
         ::Leftovers::ProcessorBuilders::Each.each_or_self(patterns) do |pat|
           case pat
           when ::Integer then ::Leftovers::Processors::PositionalArgument.new(pat, processor)
           when '*' then ::Leftovers::Processors::EachPositionalArgument.new(processor)
           when '**' then ::Leftovers::Processors::EachKeywordArgument.new(processor)
+          when /\A(\d+)\+\z/
+            ::Leftovers::Processors::EachPositionalArgumentFrom.new(pat.to_i, processor)
           when ::String, ::Hash
             ::Leftovers::ProcessorBuilders::KeywordArgument.build(pat, processor)
             # :nocov: