+++ lib/leftovers/processor_builders/transform_set.rb
@@ -7,11 +7,7 @@ module Leftovers
         def build(transforms, final_processor)
           each_builder(final_processor).each_or_self(transforms) do |transform|
             case transform
-            when ::Hash
-              next build(transform[:transforms], final_processor) if transform[:transforms]
-
-              ::Leftovers::ProcessorBuilders::TransformChain.build(transform, final_processor)
-            when ::String
+            when ::Hash, ::String
               ::Leftovers::ProcessorBuilders::TransformChain.build(transform, final_processor)
             # :nocov:
             else raise Leftovers::UnexpectedCase, "Unhandled value #{transform.inspect}"